DB migration with FlyWay, add User profile

+++ freeter/src/main/java/com/serjer/freeter/repos/MessageRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.serjer.freeter.domain.Message;
 
-public interface MessageRepo extends JpaRepository<Message, Integer> {
+public interface MessageRepo extends JpaRepository<Message, Long> {
 	
 	List<Message> findByTag(String tag);
